refa: fix directory name typo
