+++ client/src/test/resources/mybatis/embedded_in_java/main/java/com/baeldung/mybatis/mapper/AddressMapper.java
