+++ client/src/test/resources/mybatis/embedded_in_xml/main/java/org/mybatis/example/GroupMapper.java
@@ -1,0 +1,9 @@
+package org.mybatis.example;
+import org.mybatis.example.Group;
+
+/**
+ * Created by ywata on 1/4/14.
+ */
+public interface GroupMapper {
+    Group select();
+}
