+++ client/src/test/resources/mybatis/embedded_in_xml/main/java/org/mybatis/example/UserMapper.java
@@ -1,0 +1,10 @@
+package org.mybatis.example;
+import org.mybatis.example.User;
+
+/**
+ * Created by ywata on 1/4/14.
+ */
+public interface UserMapper {
+    User select();
+    int insertUser(User user);
+}
