--- client/src/test/resources/mybatis/embeded_in_xml/main/java/org/mybatis/example/UserMapper.java
@@ -1,10 +1,0 @@
-package org.mybatis.example;
-import org.mybatis.example.User;
-
-/**
- * Created by ywata on 1/4/14.
- */
-public interface UserMapper {
-    User select();
-    int insertUser(User user);
-}
