--- client/src/test/resources/mybatis/embeded_in_xml/main/java/org/mybatis/example/mybatisConfig.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE configuration PUBLIC "-//mybatis.org//DTD Config 3.0//EN" "http://mybatis.org/dtd/mybatis-3-config.dtd">
-<configuration>
-  <environments default="development">
-    <environment id="development">
-      <transactionManager type="JDBC"/>
-      <dataSource type="POOLED">
-        <property name="driver" value="com.mysql.jdbc.Driver"/>
-	<property name="url" value="jdbc:mysql://localhost:3306/test"/> 
-        <property name="username" value="root"/>
-        <property name="password" value="root"/>
-      </dataSource>
-    </environment>
-  </environments>
-  <mappers>
-    <mapper resource="User.xml"/>
-  </mappers>
-</configuration>
