--- client/src/test/resources/mybatis/embeded_in_xml/main/resources/Configuration.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?> 
-<!DOCTYPE configuration PUBLIC "-//mybatis.org//DTD Config 3.0//EN"
-"http://mybatis.org/dtd/mybatis-3-config.dtd">
-<configuration>
-    <environments default="development">
-        <environment id="development">
-            <transactionManager type="JDBC"/>
-            <dataSource type="POOLED">
-                <property name="driver" value="${driver}"/>
-                <property name="url" value="${url}"/>
-                <property name="username" value="${username}"/>
-                <property name="password" value="${password}"/>
-            </dataSource> </environment>
-    </environments>
-    <mappers>
-        <mapper resource="org/mybatis/example/User.xml"/>
-    </mappers>
-</configuration>
