--- client/src/test/resources/mybatis/embeded_in_xml/main/resources/Mapper.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="User">
-  <select id="select" parameterType="java.util.HashMap" resultType="java.util.HashMap">
-    SELECT
-      USERID,
-      USERNAME
-    FROM
-      USER
-  </select>
-  <select id="selectX" parameterType="String" resultType="String">
-     SELECT
-       USERNAME
-     FROM
-       USER
-  </select>
-  <insert
-      id="insertUser"
-      parameterType="org.mybatis.example.User"
-      flushCache="true"
-      timeout="20">
-  INSERT INTO USER (USERID, USERNAME)
-  VALUES (#{userid}, #{username})
-
-  </insert>
-</mapper>
