--- client/src/test/resources/mybatis/embeded_in_xml/main/resources/mybatisConfig.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE configuration PUBLIC "-//mybatis.org//DTD Config 3.0//EN" "http://mybatis.org/dtd/mybatis-3-config.dtd">
-<configuration>
-  <environments default="development">
-    <environment id="development">
-      <transactionManager type="JDBC"/>
-      <dataSource type="POOLED">
-        <property name="driver" value="com.mysql.jdbc.Driver"/>
-        <property name="url" value="jdbc:mysql://localhost/test"/>
-       <!-- <property name="url"  value="jdbc:mysql://opt/local/var/run/mysql55/mysqld.sock/test" /> -->
-        <property name="username" value="root"/>
-        <property name="password" value=""/>
-      </dataSource>
-    </environment>
-  </environments>
-  <mappers>
-    <mapper resource="UserMapper.xml"/>
-    <mapper resource="GroupMapper.xml"/>
-  </mappers>
-</configuration>
